deletevolume fix

+++ service/controller.go
@@ -395,15 +395,23 @@ func (s *service) DeleteVolume(
 	} else {
 		log.Info("Protocol NFS,check if volume is replicated")
 		fileAPI := gounity.NewFilesystem(unity)
-		filesystem, _ := fileAPI.FindFilesystemByID(ctx, volID)
-		replAPI := gounity.NewReplicationSession(unity)
-		session, _ := replAPI.FindReplicationSessionBySrcResourceID(ctx, filesystem.FileContent.StorageResource.ID)
-		log.Debugf("Replication session: %v", session)
-		if session != nil {
-			log.Info("Replication enabled, delete replication session")
-			err = replAPI.DeleteReplicationSessionByID(ctx, session.ReplicationSessionContent.ReplicationSessionID)
+		filesystem, err := fileAPI.FindFilesystemByID(ctx, volID)
+		if err != nil {
+			return nil, err
+		}
+		if filesystem != nil {
+			replAPI := gounity.NewReplicationSession(unity)
+			session, err := replAPI.FindReplicationSessionBySrcResourceID(ctx, filesystem.FileContent.StorageResource.ID)
 			if err != nil {
 				return nil, err
+			}
+			log.Debugf("Replication session: %v", session)
+			if session != nil {
+				log.Info("Replication enabled, delete replication session")
+				err = replAPI.DeleteReplicationSessionByID(ctx, session.ReplicationSessionContent.ReplicationSessionID)
+				if err != nil {
+					return nil, err
+				}
 			}
 		}
